more auth check

+++ app/Http/Controllers/PostsController.php
@@ -84,6 +84,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        } 
         return view('posts.edit')->with('post', $post);
     }
 
@@ -119,6 +124,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        } 
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
